refactor(crystal-ball): remove css double extension naming, it fools bundlers
